requirements.txt
django
mysqlclient
scikit-learn
tensorflow
pandas
numpy
nltk
spacy
transformers
matplotlib
seaborn
plotly
python-dotenv
passlib[bcrypt]
Pillow
mediapipe
opencv-python
anthropic
